reuseport_user: create a REUSEPORT_SOCKARRAY map and populate it.
We insert the userspace helper's very own socket into it and in the main
loop of the userspace helper we print the contents of the map.

Currently the map is still per userspace program and NOT shared. In
other words, starting two userspace helpers does not result in a merge
of map entries :-(.