reuseport_{kern,user}: Export hash bucket count in map.
This removes the need to recompile the BPF program for resizes.

Usage:
`./reuseport 0/5` will declare that there are 5 buckets globally and use the 1st one.

`./reuseport 3` will read the global count (MAX_BALANCER_COUNT if 0) and use the 2nd one.

`./reuseport` will read the global count (MAX_BALANCER_COUNT if 0) and use the 1st one.